sensor-simulator/requirements.txt
# requirements.txt
boto3
crcmod
numpy
pandas
aiomqtt
python-dotenv
s3fs
dask[distributed]
plotly